RF imports: relocate filters, gamma
leave dated backwards-compatible files that log a deprecation warning
