+++ psychopy/visual/window.py
@@ -80,7 +80,7 @@ from psychopy.core import rush
 
 reportNDroppedFrames = 5  # stop raising warning after this
 
-from psychopy.gamma import getGammaRamp, setGammaRamp, setGamma
+from psychopy.visual.gamma import getGammaRamp, setGammaRamp, setGamma
 #import pyglet.gl, pyglet.window, pyglet.image, pyglet.font, pyglet.event
 from . import shaders as _shaders
 try:
